fixed limited thread count and shared mem
